pyproject.toml
[project]
name = "django-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "channels-redis>=4.2.1",
    "channels[daphne]>=4.2.2",
    "daphne>=4.2.1",
    "django>=5.2.4",
    "django-allauth>=65.9.0",
    "redis>=6.2.0",
]
